Move SimpleCoordinates to test utils file.
