+++ glue_plotly/web/qt/tests/test_plotly.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import, division, print_function
 import pytest
 import numpy as np
 
-from glue.core import Coordinates, Data, DataCollection
+from glue.core import Data, DataCollection
 
 pytest.importorskip('qtpy')
 
@@ -13,25 +13,7 @@ from glue.viewers.histogram.qt import HistogramViewer  # noqa: E402
 from glue.viewers.profile.qt import ProfileViewer  # noqa: E402
 
 from ...export_plotly import build_plotly_call  # noqa: E402
-
-
-class SimpleCoordinates(Coordinates):
-
-    def __init__(self):
-        super().__init__(pixel_n_dim=3, world_n_dim=3)
-
-    def pixel_to_world_values(self, *args):
-        return tuple([2.0 * p for p in args])
-
-    def world_to_pixel_values(self, *args):
-        return tuple([0.5 * w for w in args])
-
-    @property
-    def axis_correlation_matrix(self):
-        matrix = np.zeros((self.world_n_dim, self.pixel_n_dim), dtype=bool)
-        matrix[2, 2] = True
-        matrix[0:2, 0:2] = True
-        return matrix
+from ....common.tests.utils import SimpleCoordinates  # noqa: E402
 
 
 class TestPlotly(object):
